requirements.txt
app==0.0.1
Faker==17.6.0
Flask==2.2.3
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
pytest==7.1.2
requests==2.28.1
WTForms==3.0.1
python-dotenv==1.0.0
